% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gsnow}
\alias{gsnow}
\title{Calculate gravity signal from snow storage}
\usage{
gsnow(snow.input)
}
\arguments{
\item{snow.input}{data.frame with columns $time, $SWE (snow water equivalent)}
}
\description{
Calculate gravity signal from snow storage
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

