% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_wettzell}
\alias{read_wettzell}
\title{Reading field measured data from BKG station Wettzell}
\usage{
read_wettzell(param = "Soil")
}
\arguments{
\item{param}{insert one of the following options to receive the newest of this data:
"Soil", "Snow", "LysiEto", "LysiWeight", "Precip", "GW", "ClimaBKG", ClimaGFZ", "Runoff"}
}
\description{
Reading measurement data from field various field equipment from the BKG Observatory Wettzell (Bavaria).
Available dataset include:

\itemize{
\item soil data (5 TDR clusters each with approx. 40 sensors)
\item snow data
\item lysimeter data (potential evaporation, weight data from the lysimeter)
\item precipition data from 4 locations ARROUND the observatory
\item groundwater data from 10 groundwater wells
\item clima data Wettzell (official data from the BKG)
\item clima data GFZ (data collected locally operated by GFZ)
\item runoff data from 4 locations arround the observatory
}
}
\details{
Data represents NON filtered raw-data directly from each individual measurement equipment loggers.
If not other stated in SI-units.
}
\examples{
readWettzell("Soil")
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

