% Generated by roxygen2 (4.0.2): do not edit by hand
\name{soilparam}
\alias{soilparam}
\title{Calculate van Genuchten parameters}
\usage{
soilparam(data.in, use.method = "VG")
}
\arguments{
\item{data.in}{Values for water retention curve. Column structure: $theta, $suction}

\item{use.method}{string of method name. possible aer VG (vanGenuchten), brook, gardner, kosugi.}
}
\description{
From a given retention curve dataset van Genuchten parameters thr, ths, alpha and n are calculated.
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

