% Generated by roxygen2 (4.0.2): do not edit by hand
\name{template}
\alias{template}
\title{test}
\usage{
template()
}
\arguments{
\item{test}{}

\item{test}{}

\item{test}{...}
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

