% Generated by roxygen2 (4.0.2): do not edit by hand
\name{theta}
\alias{theta}
\title{Calculate soil moisture content from suction potential}
\usage{
theta(suction, vgparam)
}
\arguments{
\item{suction}{Suction pressure value (in cm) to be converted to soil moisture}

\item{vgparam}{data.frame of van Genuchten parameters with the column structure: $thr, $ths, $alpha, $n}
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

