% Generated by roxygen2 (4.0.2): do not edit by hand
\name{timeshift}
\alias{timeshift}
\title{Time lag estimation via cross-correlation of two timeseries}
\usage{
timeshift(data1, data2, nmax, norm = T, stl = T, plotting = T,
  swin = 24, twin, ...)
}
\arguments{
\item{data1,data2}{Timeseries of type .zoo.}

\item{nmax}{Number of correlation values output.}

\item{norm}{logical. Use normalized data (default) or not.}

\item{stl}{logical. Use decomposed data for cross-correlation (default) or not.}

\item{plotting}{logical. Output cross-correlation plots. Default is F, own output plots are provided. Should be kept FALSE.}

\item{swin}{Time window used for seasonality estimation. Only used if stl=T.}

\item{twin}{Time window used for trend estimation. Only used if stl=T.}

\item{...}{Further parameters passed to internal functions.}
}
\description{
This function is based on standard time series decomposition and afterwards cross-correlation estimation between the two input datasets.
}
\details{
missing
}
\examples{
outside.cor = ccf.zoo(besidesBuilding$mux43_04,besidesBuilding$mux43_08,5,T,T,T,24,17521)
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}

